Tea-4925: Små justeringer på variabelnavn, flyttet tester fra TilkjentYtelseUtilsTest til SatsServiceTest
